# backend/doc/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: ProdCom-E API
  description: An e-commerce website to deliver best and authentic gadget items.
  contact:
    name: Md. Sifytul Karim
    email: sifytul09@gmail.com
...
servers:
  - description: localhost
    url: http://localhost:4000
paths:
...
  '/me':
    get:
      description: This will retrieve the profile data based on the provided token.
      tags: [User]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User data
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      name:
                        type: string
                        example: Sifytul Karim
                      email:
                        type: string
                        example: sifytul@example.com
                      avatar:
                        type: string
                        example: s3.avatar.jpeg
                      role:
                        type: string
                        example: user
                      userSince:
                        type: string
                        example: 2023-10-13T09:25:14.836Z
  '/me/password/update':
    patch:
      tags: [User]
      security:
        - bearerAuth: []
      description: This route will update the password
      requestBody:
        description: take old and new password
        content:
          'application/json':
            schema:
              type: object
              properties:
                oldPassword:
                  type: string
                  example: password123
                newPassword:
                  type: string
                  example: newpassword123
                  minLength: 8
                wantToLogOutFromOtherDevices:
                  type: boolean
                  example: false
              required: [oldPassword, newPassword]
      responses:
        '200':
          description: success response
          content:
            'application/json':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Password successfully changed.
        '400':
          description: wrong old password or new password validation error
          content:
            'application/json':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: wrong password
        '401':
          $ref: '#/components/responses/401'
  '/me/update':
    patch:
      description: This route will update the name or avatar of the user.
      tags: [User]
      security:
        - bearerAuth: []
      requestBody:
        content:
          'application/json':
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: New Name
                avatar:
                  type: string
                  example: s3.fdfjdlfr.jpg
      responses:
        '204':
          description: success response
          content:
            'application/json':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: requested field successfully changed.
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
  '/me/update/avatar':
    post:
      tags: [User]
      security:
        - bearerAuth: []
      description: This route will update the avatar of the user.
      requestBody:
        content:
          'multipart/form-data':
            schema:
              type: object
              properties:
                avatar:
                  description: only jpg, jpeg, png and file less than 1mb is allowed
                  type: string
                  format: binary
      responses:
        '204':
          description: success response
          content:
            'application/json':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: avatar successfully updated.
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '500':
          $ref: '#/components/responses/500'
  # User - Admin endpoints
  '/admin/users':
    get:
      tags: [User]
      security:
        - bearerAuth: []
      description: Retrieve all users list. Admin only
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/sordTypeParam'
        - $ref: '#/components/parameters/sortByParam'
        - $ref: '#/components/parameters/searchParam'

      responses:
        '200':
          description: list of users
          content:
            'application/json':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  users:
                    type: array
                    items:
                      type: object
                      $ref: '#/components/schemas/User'

        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
    post:
      tags: [User]
      description: This route will create a user with provided payload by admin
      security:
        - bearerAuth: []

      requestBody:
        content:
          'application/json':
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: 'Sifytul Karim'
                email:
                  type: string
                  example: sifytul@example.com
                  format: email
                password:
                  type: string
                  example: password123
                  format: password
              required: [name, email, password]

      responses:
        '201':
          description: registered a new user
          content:
            'application/json':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '400':
          description: Bad request
          content:
            'application/json':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                          example: email
                        message:
                          type: string
                          example: Field can't be empty
        '409':
          $ref: '#/components/responses/409'
  '/admin/users/{id}':
    get:
      tags: [User]
      security:
        - bearerAuth: []
      description: Retrieve a single user's data
      parameters:
        - in: path
          name: id
          required: true
          description: user id
          schema:
            type: integer
            example: 2
      responses:
        '200':
          description: detailes of a single user
          content:
            'application/json':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
    patch:
      tags: [User]
      security:
        - bearerAuth: []
      description: update any user data
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            example: 2
      requestBody:
        content:
          'application/json':
            schema:
              type: object
              properties:
                role:
                  type: string
                  enum: [user, admin]
                  example: user
      responses:
        '200':
          description: success response
          content:
            'application/json':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: role changed successfully

        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
    delete:
      tags: [User]
      security:
        - bearerAuth: []
      description: This route will delete a user
      parameters:
        - in: path
          name: id
          required: true
          description: user id
          schema:
            type: integer
            example: 2
      responses:
        '200':
          description: success delete response
          content:
            'application/json':
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: successfully deleted user with id 2
        '400':
          description: not found error
          content:
            'application/json':
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User not found
                  error:
                    type: string
                    example: Bad Request
                  statusCode:
                    type: integer
                    example: 400
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'

  # Products Endpoints
...
components:
  responses:
    '400':
      description: Bad request
      content:
        'application/json':
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              errors:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      example: required field
                    message:
                      type: string
                      example: Field can't be empty
    '409':
      description: Conflict
      content:
        'application/json':
          schema:
            type: object
            properties:
              status:
                type: integer
                example: 409
              message:
                type: string
                example: duplicate email
    '401':
      description: Unauthorized
      content:
        'application/json':
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Please sign in to perform the action.
    '403':
      description: Forbidden
      content:
        'application/json':
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: You are not authorized to see these resources
    '500':
      description: Server error
      content:
        'application/json':
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              message:
                type: string
                example: Internal server Error
...
  parameters:
...
    pageParam:
      in: query
      name: page
      required: false
      description: page no
      schema:
        type: integer
        format: int32
        example: 2
    limitParam:
      in: query
      name: limit
      required: false
      description: number of users in one page
      schema:
        type: integer
        format: int32
        example: 10
    sordTypeParam:
      in: query
      name: sort_type
      description: how the data will be sorted (asc or desc)
      schema:
        type: string
        enum: [asc, desc]
        example: desc
    sortByParam:
      in: query
      name: sort_by
      description: by which field the users will be sorted(by createdAt or name)
      schema:
        type: string
        enum: [created_at, name]
        example: created_at
    searchParam:
      in: query
      name: search
      description: search term
      schema:
        type: string
        example: sifytul
...
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 434
        name:
          type: string
          example: Sifytul Karim
        email:
          type: string
          example: sifytul@example.com
        address:
          $ref: '#/components/schemas/Address'
        contact:
          $ref: '#/components/schemas/ContactInfo'
        avatar:
          type: string
          example: s3.avatar.jpeg
        links:
          type: object
          properties:
            self:
              type: string
              example: /admin/users/3
        orderDetails:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
                example: 3
              orderDate:
                type: string
                example: 2023-09-26 03:03:00
              status:
                type: string
                example: processing
              itemsPrice:
                type: integer
                example: 345.34
              links:
                type: object
                properties:
                  self:
                    type: string
                    example: /orders/3
...
    Address:
      type: object
      properties:
        id:
          type: integer
          example: 3
        address:
          type: string
          example: '204, sholoshohor'
        city:
          type: string
          example: chittagong
        state:
          type: string
          example: chittagong
        pinCode:
          type: integer
          example: 4333
    ContactInfo:
      type: object
      properties:
        phone1:
          type: string
          example: +8801000000000
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT